pyproject.toml
[tool.poetry]
name = "language-simulator-backend"
version = "0.1.0"
description = ""
authors = ["Akim Abdulaev <a.abdulaev@idaproject.com>"]
readme = "README.md"
packages = [{include = "language_simulator_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
python-multipart = "^0.0.6"
email-validator = "^1.3.1"
pydantic = "^1.10.6"
emails = "^0.6"
python-jose = "^3.3.0"
passlib = "^1.7.4"
alembic = "^1.10.2"
uvicorn = "^0.21.0"
fastapi = "^0.94.0"
dependency-injector = "^4.41.0"
asyncpg = "^0.27.0"
greenlet = "^2.0.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.20.3"
factory-boy = "^3.2.1"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
